space/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.4.0
gradio==4.44.1
unsloth==2026.5.2
peft>=0.13.0
transformers>=4.46.0
accelerate>=1.0.0
bitsandbytes>=0.45.0
sentencepiece>=0.2.0
huggingface_hub>=0.26.0
spaces>=0.30.0
